Add Thrift 0.9.3
This commit adds package Thrift 0.9.3 to CCI.

The previous test_package Calculator application was dependent on code generated
using the Thrift 0.13.0 compiler. As such, the Calculator application cannot support
Thrift 0.13.0 and 0.9.3. Therefore, the recipes for 0.9.3 and 0.13.0 are in individual folders.

This could likely be resolved so that the Thrift recipe was dependent on a Conan package containing
the Thrift compiler and the test_package conanfile.py generated the C++ code using the correct Thrift
compiler version